refactor: move useFeature to schema.ts
