+++ app/components/Form/ApplicationForm.tsx
@@ -23,37 +23,7 @@ const ApplicationForm: React.FC<Props> = ({
   const router = useRouter();
   const [updateApplication] = useUpdateApplicationMutation();
 
-  // Check if development form is enabled in growthbook and pass to schema
-
-  const formAlternateContact = useFeature('form-alternate-contact').value;
-  const formAuthorizedContact = useFeature('form-authorized-contact').value;
-  const formBudgetDetails = useFeature('form-budget-details').value;
-  const formContactInformation = useFeature('form-contact-information').value;
-  const formExistingNetworkCoverage = useFeature(
-    'form-existing-network-coverage'
-  ).value;
-  const formOrganizationLocation = useFeature(
-    'form-organization-location'
-  ).value;
-  const formOrganizationProfile = useFeature('form-organization-profile').value;
-  const formProjectArea = useFeature('form-project-area').value;
-  const formProjectInformation = useFeature('form-project-information').value;
-
-  const featureFlagsForm = {
-    formAlternateContact,
-    formAuthorizedContact,
-    formBudgetDetails,
-    formContactInformation,
-    formExistingNetworkCoverage,
-    formOrganizationLocation,
-    formOrganizationProfile,
-    formProjectArea,
-    formProjectInformation,
-  };
-
-  const subschemaArray = schemaToSubschemasArray(
-    schema(featureFlagsForm) as object
-  );
+  const subschemaArray = schemaToSubschemasArray(schema() as object);
 
   if (subschemaArray.length < pageNumber) {
     // Todo: proper 404
