+++ app/formSchema/schema.ts
@@ -1,3 +1,4 @@
+import { useFeature } from '@growthbook/growthbook-react';
 import {
   alternateContact,
   authorizedContact,
@@ -11,20 +12,22 @@ import {
   projectFunding,
 } from './pages';
 
-const useSchema = (featureFlagsForm: any) => {
-  // featureFlagsForm is passed and enables development pages if set
-  const {
-    formAlternateContact,
-    formAuthorizedContact,
-    formBudgetDetails,
-    formContactInformation,
-    formExistingNetworkCoverage,
-    formOrganizationLocation,
-    formOrganizationProfile,
-    formProjectArea,
-    formProjectInformation,
-    formProjectFunding,
-  } = featureFlagsForm;
+const useSchema = () => {
+  // Check if development form is enabled in growthbook and pass to schema
+  const formAlternateContact = useFeature('form-alternate-contact').value;
+  const formAuthorizedContact = useFeature('form-authorized-contact').value;
+  const formBudgetDetails = useFeature('form-budget-details').value;
+  const formContactInformation = useFeature('form-contact-information').value;
+  const formExistingNetworkCoverage = useFeature(
+    'form-existing-network-coverage'
+  ).value;
+  const formOrganizationLocation = useFeature(
+    'form-organization-location'
+  ).value;
+  const formOrganizationProfile = useFeature('form-organization-profile').value;
+  const formProjectArea = useFeature('form-project-area').value;
+  const formProjectInformation = useFeature('form-project-information').value;
+  const formProjectFunding = useFeature('form-project-funding').value;
 
   const schema = {
     type: 'object',
